Experimental (#14)
* limit rectilinear merge

* message box

* fix iostream include

* rework model loading 1

* obj loading rework part 2

* model loading rework part 3

* triangle mesh generator rework + cuda sm fix

* convert triangle mesh face to ivec3

* unified model vertex index separate export

* textured asset

* triangle mesh v123 fix

* triangle mesh v123 fix 2

* fix pi + update travis eigen

* fix travis path + core api to vertexlist()

* objloader cleanup

* remove obj asset loader

* remove assetloader

* remove model.h modelloader.h

* move animation to core + use sdl2s findconfig

* add animation sample

* sdl cmake target

* minor cmake stuff

* update sdl2 cmake

* travis vcpk test

* even more sdl changes + windows travis

* fix some warnings

* windows fix

* remove redundant code

* rename model loader files

* a few more renames

* install more windows vcpkg packages

* remove some compiler flags for windows benchmarks

* mini fix

* travis test

* travis test2

* add submodule eigen

* reworked samples

* travis ;

* travis check

* travis again

* assimp not required anymore

* use gtest config

* gtest cmake + windows test fixes

* epsilon test

* no win32 poisson

* rendering stuff

+++ src/saiga/core/geometry/RectangularDecomposition.cpp
@@ -63,7 +63,7 @@ RectangleList DecomposeRowMerge(ArrayView<const ivec3> points)
     Rect current = Rect(copy.front());
 
 
-    for (int i = 1; i < copy.size(); ++i)
+    for (int i = 1; i < (int)copy.size(); ++i)
     {
         auto index = copy[i];
 
@@ -140,7 +140,7 @@ RectangleList DecomposeOctTree(ArrayView<const ivec3> points, float merge_factor
         int range_begin = 0;
         int range_end   = 0;
 
-        for (int i = 0; i < copy.size(); ++i)
+        for (int i = 0; i < (int)copy.size(); ++i)
         {
             auto& prev = copy[range_begin];
             auto& curr = copy[i];
@@ -152,7 +152,7 @@ RectangleList DecomposeOctTree(ArrayView<const ivec3> points, float merge_factor
 
             //            std::cout << (prev.first & mask) << " " << (curr.first & mask) << std::endl;
 
-            if ((prev.first & mask) == (curr.first & mask) && i != copy.size() - 1)
+            if ((prev.first & mask) == (curr.first & mask) && i != (int)copy.size() - 1)
             {
                 //                std::cout << "merge " << bit << " " << prev.second << " with " << curr.second <<
                 //                std::endl;
@@ -164,7 +164,7 @@ RectangleList DecomposeOctTree(ArrayView<const ivec3> points, float merge_factor
             }
             else
             {
-                if (i == copy.size() - 1)
+                if (i == (int)copy.size() - 1)
                 {
                     range_end++;
                 }
